✨ feat: 인증 설정 단계 API 수정 및 인증 옵션 조회 API 추가

+++ src/main/java/com/oss/saber/dto/VerificationLinkSettingRequest.java
@@ -12,5 +12,5 @@ import java.util.List;
 public class VerificationLinkSettingRequest {
     private Integer limitedMinutes;
     private String customRequests;
-    private List<Integer> VerificationMethods;
+    private List<Long> VerificationMethods;
 }
